refactor: move utility functions from utils to helpers/utils for better organization
