+++ nodejs/test/bulkPulling/schemaless.test.ts
@@ -2,7 +2,7 @@ import { WebSocketConnectionPool } from "../../src/client/wsConnectorPool";
 import { WSConfig } from "../../src/common/config";
 import { Precision, SchemalessProto } from "../../src/sql/wsProto";
 import { WsSql } from "../../src/sql/wsSql";
-import { testPassword, testUsername } from "../utils";
+import { testPassword, testUsername } from "../helpers/utils";
 
 let dns = "ws://localhost:6041";
