+++ nodejs/test/bulkPulling/wsProxy.failover.test.ts
@@ -1,7 +1,7 @@
 import { WebSocketConnectionPool } from "../../src/client/wsConnectorPool";
 import { WSConfig } from "../../src/common/config";
 import { WsSql } from "../../src/sql/wsSql";
-import { testPassword, testUsername } from "../utils";
+import { testPassword, testUsername } from "../helpers/utils";
 import { WsProxy, WsProxyEvent } from "../helpers/wsProxy";
 
 function parseBinaryAction(rawData: Buffer | string): bigint | null {
